backend/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# LangChain and AI
langchain==1.2.6
langchain-core==1.2.7
langchain-classic==1.0.1
langchain-openai==1.1.7
langchain-chroma>=0.1.2
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
openai==2.15.0
chromadb>=0.4.0

# NLP and sentiment analysis
textblob==0.17.1
vaderSentiment==3.3.2
nltk==3.8.1

# Data validation
pydantic==2.12.5
pydantic-settings==2.12.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
